backend/requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12.0
django-cors-headers>=3.7.0
gunicorn>=20.1.0
psycopg2-binary>=2.9.1
dj-database-url>=0.5.0
bcrypt>=4.3.0